+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,6 +1,7 @@
 Version 6.8.2
 =============
 
+* 2021-03-19 Streamline tests (asolntsev)
 * 2021-03-19 Updated Guice dependency to 5.0.1 (asolntsev)
 * 2021-03-19 Updated dependencies: slf4j:1.7.30, jackson:2.12.2, httpclient:4.5.13, commons-codec:1.15 (asolntsev)
 * 2021-03-19 Updated test dependencies: junit:4.13.2, mockito:3.8.0, powermock:2.0.9, greenmail:1.6.2, fluentlenium:3.10.1 (asolntsev)