+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,6 +1,8 @@
 Version 6.8.2
 =============
 
+* 2021-12-24 Fixed NullPointerException on "Message::getWithDefault" when value and default value are both null
+  (thibaultmeyer)
 * 2021-12-17 Updated Flyway dependency to 8.2.2 (asolntsev, thibaultmeyer)
 * 2021-03-19 Streamline tests (asolntsev)
 * 2021-03-19 Updated Guice dependency to 5.0.1 (asolntsev)